# api/openapi.yaml
openapi: 3.0.3
info:
  title: Deployment Schedule API
  description: API for managing deployment schedules
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
  /auth/google/login:
    get:
      summary: Initiate Google OAuth login
      operationId: googleLogin
      tags:
        - authentication
      responses:
        '307':
          description: Redirect to Google OAuth
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/google/callback:
    get:
      summary: Google OAuth callback
      operationId: googleCallback
      tags:
        - authentication
      parameters:
        - name: code
          in: query
          required: true
          description: Authorization code from Google
          schema:
            type: string
        - name: state
          in: query
          required: true
          description: State parameter for CSRF protection
          schema:
            type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                  - user
                properties:
                  token:
                    type: string
                    description: JWT token for API authentication
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: Refresh JWT token
      operationId: refreshToken
      tags:
        - authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                properties:
                  token:
                    type: string
                    description: New JWT token
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Google OAuth login
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        name:
          type: string
          description: User display name
        role:
          type: string
          enum: [viewer, deployer, admin]
          description: User role
        lastLoginAt:
          type: string
          format: date-time
          description: When the user last logged in
        createdAt:
          type: string
          format: date-time
          description: When the user was created
        updatedAt:
          type: string
          format: date-time
          description: When the user was last updated
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details